refactor and move functions
