+++ lambdas/routes/users/main.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
-from .authentication import router as authentication_router
+from lambdas.routes.authentication.main import router as authentication_router
 from ...db.main import get_db_session, User
 from ...dtos.users import UpdateRequestDTO
